requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
